Move markdown-related classes to own module
`formatters.py` is a big file containing a mixture of:
- short functions which take mostly taker a single string argument and
  transform it somehow
- big class definitions which inherit from Mistune (the library we use
  to parse Markdown)

This commit splits the file in two, putting the latter into a new file
called `markdown.py`. It feels like the code is easier to navigate this
way.

By keeping the functions which take a string and apply the Markdown
formatting in `formatters.py` this should be non-disruptive change.