Make links without http clickable in text messages and alerts
Phones render URLs in text messages and emergency alerts even if those
URLs don’t start with the protocol (http:// or https://)

Although we’d prefer that people use ‘full URLs, starting with https://’
our previews should be as faithful as possible to what recipients are
going to see on their phones.

This will also help people spot when they are accidentally making things
clickable.For example when missing out a space after a full stop.