Be more sophisticated with URLs ending in ) or .
Often when text contains a URL ending in a closing bracket or a full
stop that punctuation is part of the text, not part of the URL.

Divining this with a regular expression is just too complicated.
Instead this commit uses regular expressions to match everything that
might be part of the URL, then does some extra processing to work out
whether certain trailing punctuation should be included in the link,
or printed after it.