Bump PyPDF2 to version 2.0.0
This is happening automatically with the rebuild of this library,
but we're seeing failures due to one of the breaking changes [^1].
I've also fixed all the deprecation warnings - the warnings contain
the alternate code to use in place of what we had.

I can't see any functionality changes in the changelog so I think
this is safe to distribute. I've locked the version to v2+ to ensure
all apps consistently pick it up in their requirements.

[^1]: https://raw.githubusercontent.com/py-pdf/PyPDF2/main/CHANGELOG