Fix auto linking domains in email addresses
Email addresses have two things which could be interpreted as URLs which
should be made in to links:

1. the domain name itself, eg @example.com
2. the local part when it contains a dot, eg firstname.lastname

Either the whole email address should be linked or none of it should be
linked. This commit preserves the previous behaviour where no part of
the email address was linked.

It doesn this by modifying the regular expression using word boundaries
(`\b`) and negative lookahead/lookbehinds to sniff out when something
that might be a link starts or ends with an `@`, suggesting it is part
of an email address.