Fix parts of email addresses wrongly being linked
We fixed the domain name being linked in email addresses when the
address was in the format `test@example.com`.

However when the address was in the format `test@test.example.com`
`example.com` was still being made into a link because the `.` between
`test` and `example` is seen as a word boundary.

This commit fixes that by adding `.` to the negative lookbehind. This
means anything which looks like a URL but starts with `.` or `@` will
not be made into a link.