Fix HTML escaping in `Field` personalisation keys
`Field.replaced` sanitises the content, includes the names of any
placeholders. It then passes the santised placeholder names into
`Field.get_replacement` which tries to replace the placeholder with its
respective value from the dictionary in `Field.values`.

The setter method of `Field.values` does not sanitise the keys before
storing the original dictionary.

This meant that placeholders with special characters (for example `&`)
would not work because the code was trying to do
`Field.values['date&amp;time']` where `Field.values` was a dictionary
with the keys `'date&time'`.

This commit changes the setter method to sanitise the keys before
storing the original dictionary. This fixes the bug, because the keys
are consisitently sanitised everywhere.