Treat narrow no break space like non breaking space when normalising and sanitising
This is a new whitespace character we didn’t know about. It appears to
perform the same function as a non-breaking space (i.e. preventing a
line break between words when used instead of a normal space).

If we don’t specifically account for it then it gets treated like any
other unusual unicode character. This means when a user tries copies
some text into their template which uses this character they see this
error:

> 'You cannot use  in text messages. It will not show up properly on
> everyone's phones.'

…which is unhelpful because it’s not obvious that the space between
‘use’ and ‘in’ is what’s being referred to.

Only a change in `sanitise_text.py` is required to fix this, but it
seems reasonable to change it in `formatters.py` too where we use a
different method to normalise the same set of whitespace characters.