Disable logging to file when we're running on ECS
This depends on a new environment variable `NOTIFY_RUNTIME_PLATFORM`
which should denote the platform we're running on.

If the value of this variable is `ecs` then we don't attempt to create
the `NOTIFY_LOG_PATH` and we do not append a file handler to our logger.

This also simplifies the `get_handlers` function to reduce the
nestedness of `if` statements.