Remove `Template.encoding`
We added the `encoding` argument to `Template` when we started counting
the length of template content:
https://github.com/alphagov/notifications-utils/pull/25/commits/a9b09d841b0abf928c145225c279b0cc7ae217df#diff-01b6821e190538339d22c9858b938ac4ce299a4b29d914eb81f9acad86ce0228R13-R29

We needed to do this because determined the length by encoding the
template to bytes then counting how many bytes there were. This meant
we needed to know what encoding the text was in otherwise we might
miscount.

Practically we always deal with text in UTF8, so this property was
hard coded on the class in
https://github.com/alphagov/notifications-utils/commit/9653f28e0448c8a20b97a4097f5dd14c27d5eb3c#diff-01b6821e190538339d22c9858b938ac4ce299a4b29d914eb81f9acad86ce0228R13-R26

We stopped counting bytes as a way of determining template length here,
which meant we stopped referring to `Template.encoding` anywhere:
https://github.com/alphagov/notifications-utils/commit/aae1c82806fa6f5067d1dea4578648a0b5fd7567#diff-01b6821e190538339d22c9858b938ac4ce299a4b29d914eb81f9acad86ce0228R181

However we didn’t remove it at that time – this commit now removes it.

This is technically a breaking change but there isn’t any code actually
using the property as far as I can tell.