replace orderedset with ordered-set
orderedset is abandoned. ordered-set is marginally less abandoned.
python's dict is implicitly ordered in python 3.7 and higher so we may
be able to create our own set class that secretly just has a dict under
the hood - but that might have a bunch of complications (mostly around
method names, overriding set operands like |, and comparison to regular
sets)