Pin pyproj below 3.4.1
We started to get a lot of test failures in utils after the release of
pyproj 3.4.1. We pulled in the latest release automatically for tests
directly on the utils repo.

Let's pin to below that version for now, so that tests can pass. We
should investigate further:

1) What's changed
2) Whether the 'issue' will be fixed in pyproj or if we should look at
updating our use of the lib to become compatible with what's changed.