Make `get_offset_days` work for arbitrary number of days
Previously this function was just a helper for getting the next or
previous working day, meaning it only ever expected to get an offset of
1 or -1 days.

Because we now want to use it to walk back n working days, this commit
modifies it to work back or forwards one day at a time, not counting
non-working days.

It then adds some helper functions to mask the internals of how we
differentiate between DVLA and Royal Mail working days.