add lock to redis client
slight niggle in that we want to return a rich object, so we need to
make a stub class for this when redis is not enabled (such as locally).

the stub class reimplements all the methods of redis.lock.Lock,
including marking the lock as locked when you call acquire or enter the
context manager. But other than that it's very simple and stupid.