Normalise BFPO addresses
Add PostalAddress support for BFPO addresses.

There's quite a lot of potential variation in how BFPO addresses can be
supplied, but a relatively clear format for how they should be output:

https://www.gov.uk/bfpo/how-to-address-bfpo-mail

At the moment we require postcodes to be supplied for all addresses, and
this should be the last line. But BFPO addresses expect, for postage,
the last line to be a BFPO number. This adds support for accepting that
format.

If a BFPO postcode is used without a BFPO number line, it won't be
detected as a BFPO address or handled in any special way, but it should
still get delivered properly by Royal Mail. We just won't be able to
mark it up nicely for DVLA to help them sort it automatically.