Include country for normalised BFPO
Although we consider this an invalid address, we should still include
the country in the output as it was provided coming in.

Some apps, eg admin, process the address multiple times and so if we
drop off the country, the latter processing steps won't consider the
address invalid. We may want to fix that separately, but let's be safe
on this side.