Add classes to mark if pages are odd or even
This pull request introduced a visual bug due to
its use of `nth-of-type` assuming
`.letter:nth-of-type(even)` would affect
even-numbered elements of type `div.letter` when
it actually matches even-numbered `div` elements,
filtered by whether they have the `letter` class
or not.

https://github.com/alphagov/notifications-admin/pull/4648

You will be able to match even-numbered pages in
CSS in future (see
https://css-tricks.com/css-nth-of-class-selector/)
but until then, we can use Jinja to add a specific
class for this instead. I've added one for both
even and odd as a convenience.

Uses BEM syntax to indicate a state of page. There
is currently no CSS in admin targeting this class
so this can be merged with no impact.