Add support for Chinese character sets in letters
The fonts we use in letters support all character sets we have tested,
except for Chinese.

This commit adds WenQuanYi Zen Hei[1] to our font stack. This is a font
which:
- is freely installable in our Docker images
- has glyphs for Chinese character sets

The way CSS works is that if the first font in the stack doesn’t have
glyphs for a particular character it will fall back to the next font,
then the next font, etc.

1. https://en.wikipedia.org/wiki/WenQuanYi