Add `class="page--first"` and `class="page--last"`
…to the first and last pages (respectively) of letters rendered with
`LetterImageTemplate`.

This will let us target them with CSS which we can’t always do at the
moment because of the limitations of selectors like `:first-child` and
`:first-of-type`

Extends the approach taken in https://github.com/alphagov/notifications-utils/pull/1025

This will let us fix a bug where at the moment all letter pages have
`0` bottom margin if the attach pages button is visible, whereas this
was supposed to apply to the last page only.