Remove auto-log formatting and standardise format
Removes the custom functionality to automatically try to inject
variables into the log string. A whole host of our log messages were
erroring out on this functionality because the variable didn't exist, or
because we were logging out dictionaries which "".format() was trying,
and failing, to handle.

Now that we've integrated Sentry, let's standardise on the "old-style"
logging format, which looks something like this:

logger.info("Response %s took %.4f to generate", response.body,
duration)

Python's logging system will automatically interpolate the message with
the provided arguments, and sentry will group any error-level log
messages as a single issue (the way we used to do things - like
`logger.error(f"something went wrong: {exc}") would generate a unique
sentry issue if exc generated a unique string for each instance of the
error).

This is a breaking change: when pulling into any of the apps we should
go through and systematically update all of the log messages to the
"old-style" %s format. We can enforce this with the "G" rule in Ruff.