Allow zendesk ticket creation without notify type
At the moment all support tickets are raised with a notify ticket type
assigned immediately (generally, non-technical). This is often
misleading as we get technical queries through this channel as well. We
can have a clearer triage process if tickets come in with no type, and
then the first person to look at the ticket can assign it to either
non-technical (product/design/research/general support) or technical
(requiring dev input).