Move `postage` argument to base class
At the moment we are forced to use `LetterImageTemplate` in a few places
in the admin app because it’s the only one that has a `postage`
attribute. The code would be less complex if we could use
`LetterPreviewTemplate` everywhere.

Moving the handling of the `postage` argument to the base class gets the
classes a bit closer to being polymorphic. This means that any subclass
(including `LetterImageTemplate` and `LetterPreviewTemplate`) will
always have a `postage` attribute.