Make `postage` a setter
It’s a bit cleaner to wrap this up in a separate setter method, rather
than having a lot of logic cluttering up the `__init__` method.

It also means that if an external consumer of the class ever tries to
set the postage the same type checking of the values will still be
applied.