Only require `image_url` to render `LetterImageTemplate`
At the moment we we require the `image_url` parameter whenever we
instantiate a `LetterImageTemplate`.

However the `image_url` is only needed when rendering the template (so
it knows where to find template preview).

This commit defers raising an exception until the consumer of the class
tries to render a `LetterImageTemplate` without specifying `image_url`.

This means we won’t need to pass in a dummy `image_url` where we are
doing other stuff with the template (like looking at the placeholders).