Make `LetterImageTemplate.page_count` a property
This means that it can be overidden by subclasses.

Previously if a subclass tried to implement a `page_count` property then
it would cause an exception where the `__init__` method is trying to
assign to `page_count`.

Letting subclasses define their own `page_count` will mean there’s a
nice place to put logic about how many pages are in a templated letter
with an attachment, and on what page the attachment starts.