Allow subheadings via ## markdown
Previously we collapsed all level-2 headings (with ##) down to basic
paragraphs. Our users want more formatting options and this is something
we can offer. We don't want to offer bold because we know that it can be
overused and lead to worse quality content, but subheadings may allow
services to better structure content and still draw attention to
important things.