Only require `page_count` to render (not instantiate) `LetterImageTemplate`
At the moment we we require the `page_count` parameter whenever we
instantiate a `LetterImageTemplate`.

However the `page_count` is only needed when rendering the template (so
it knows how many `<img>` tags to spit out).

This commit defers raising an exception until the consumer of the class
tries to render a `LetterImageTemplate` without specifying `page_count`.

This means we won’t need to pass in a dummy `page_count` where we are
not sure about the `page_count` until later on.