Remove postage argument from letter templates
There are a few places where an app overrides a template’s postage (for
example by setting it to `None` when the letter has failed validation).

I think it’s better to make it explicit that this is an override, rather
than leaving implementers to wonder whether `template["postage"]` or
`postage=` will take priority when initialising the class.