Extend request cache expiry from 7 to 28 days
According to the Redis metrics from the PaaS the most we’ve used in the
last 30 days is about 20% of the 568mb provided by our plan.

The longer we keep things in the cache, the more likely they are to be
served from it, rather than having to go to the origin  (the API).
Serving things from the cache is generally faster than serving them from
the origin.

The worst case scenario is that by quadrupling the expiry time we
quadruple the number of items in the cache. This would still only use
80% of the available memory, which still gives some headroom.

The worst case scenario is unlikely, because many of the things we cache
at the moment will be accessed frequently. This means that they will get
re-accessed and re-cached soon after they expire.

Most of the performance benefit here will come for items that are
infrequently accessed. For example users who don’t log in very often or
templates which aren’t sent very often. So while it may be a small
improvement it’s worth doing – we are paying for the storage either way.