Add mappings for international phone countries
If we want to display the correct names for countries we need to be able
to map from the source data to what we consider the canonical names.

This commit adds data to our countries list so that we can do that.

Whether or not this is used for international SMS, it’s always good to
make these mappings more robust and comprehensive for international
letters.