Style long QR code placeholder text better
When a QR code contains a placeholder we render a ‘placeholder’ box
instead, which has a dashed border.

If the text of the placeholder or its prefix/suffix is long, then this
causes the text to wrap inside the placeholder box. Because the
placeholder box is a relatively narrow fixed width this:
- happens easily
- does not look very good
- makes the text harder to read

This commit adds some extra HTML elements which allow us to style the
box and its contents separately.

This way the contents can overflow the box without looking messy. The
width of the text is therefore limited to the width of the column, as it
is for other URLs and placeholders.

Ideally this would use vertical-align: middle to centre the text, even
when it spans multiple lines. But Weasyprint doesn’t seem to support
this. Instead we use using a negative margin to vertically centre one
line of text, which will have to do.