Raise error if QR code too long
We want to make sure that any QR codes generated for letters are of a
usable pixel density/print quality when actually on the page. Given that
we are printing QR codes at a fixed size on the page, the more data the
QR code contains, the smaller each pixel will be printed. Smaller pixels
can make the QR code harder to scan and will require higher print
quality/resolution.

For now we are happy to balance print quality and data volumes at about
~500 bytes. The next resolution bump for QR codes generated with
'medium' error correction (15%) happens at 505 bytes.

Because QR codes for letters can have both static and dynamic
(personalised) content we need to validate the amount of QR code in two
places:

For fully static QR code data, we can validate it when the template is
being created/updated. We add a `has_qr_code_with_too_much_data`
property that renders the template content with a new validating
renderer which throws errors if too much QR code data is passed in.

For QR codes with personalisation, we can only validate the length when
the personalisation is provided, which can happen in a few different
ways.

- on the admin app with a one-off send
- on the admin app with a CSV upload
- on the API with a one-off send