Use a simpler syntax for QR codes in letters
Explaining our current syntax for QR codes is pretty gnarly, especially
when it comes to using placeholders as all or part of the URL, for
example
> [qr](https://www.example.com?reference=((placeholder)))

If I saw triple closing brackets like that in some code I was writing I
would consider if there was a better way tofactor the code to make it
easier to read. And this example isn’t full-blown code, it’s our
templating.

We initially went with this syntax because it’s consistent with how
Markdown links work. However it also differs from Markdown links in that
the label part (the [qr]) is fixed and can’t be changed.

Markdown links do have ther advantage of being inline, in other words
they can be inserted in the middle of a paragraph. But this isn’t
something we’d ever want to do because we always display the QR codes
much larger than a single line, with padding around them.

Re-using Markdown links also has the disadvantage that our Markdown
parser will be expecting URLs between the parentheses. In the future we
users might want to include arbitrary data in their QR codes (for
example some kind of encrypted blob). If this data included spaces or
parentheses then the parser which is looking for a URL might struggle.

What I’m proposing here is an alternative syntax which has no brackets:
> QR: https://www.example.com?reference=((placeholder))

I think this is:
- easier to read
- easier to explain in guidance
- more future proof in terms of what data can go in it (basically
  anything except newlines)

This commit doesn’t remove the old way of doing things yet – we can do
that separately if we decide for sure that this is the way we want to
go.