Remove the old syntax for QR codes in letters
We’ve introduced a new, friendlier syntax for QR codes in letters, which
looks like
```
QR: https://www.example.com
```

The old one looks like:
```
[QR](https://www.example.com)
```

We should remove the old one so there’s only one way to do things.

We should remove it now (before we launch the feature) so that we don’t
have to worry about it being a breaking change for teams using the old
syntax.

This also includes making the regex for the new syntax a bit more
flexible so it can handle leading whitespace.