Return QR code validation error from check
The exception raised by the contains useful information about the
problem at hand, such as the data passed into the QR code, and the
actual & max length of QR codes.

Let's pass this back instead of lose the information by only returning
True/False. Then we can potentially display the bad QR code data to the
user if it's helpful.