Add original protocol to `<strong>` in letter links
We remove the protocol from URLs in letters

If we want to reconstruct the full URL later down the line, we need to
store the original protocol somewhere.

HTML allows for arbitrary attributes to be created as long as they are
prefixed with `data-`. So this commit introduces a new attribute on the
`<strong>` tag which stores the original protocol for later use.