Remove strong tags from links in QR codes
We automatically make URLs in letters bold by adding a `<strong>` tag
around them. We also remove the protocol (for example `https://`) from
the URL.

This happens before we look for our QR code syntax by hijacking the
`paragraph()` method.

This means that URLs going into QR codes end up with a small bit of
markup in them. This prevents them from going to the original URL when
scanned.

This commit:
- removes the `<strong>` tag
- restores the protocol
- preserving any placeholders

Ideally we wouldn’t have to do and then undo this work, but I can’t see
a way to either:
- selectively not add the `<strong>` tag if the URL is inside a QR code
- run the QR code detection before running the autolink code