Remove line and paragraph separators from SMS
These are obscure unicode characters which are used to indicate that a
new line should be treated as a line break or a new paragraph.

Because they are not in the GSM-7 character set we can’t send them
without incurring double-charging. For this reason we validate them out
at the time of template editing, and the user gets this unhelpful
error message:

> You cannot use  in text messages. It will not show up properly on
> everyone’s phones

This commit adds these characters to the list of whitespace we remove
and sanitise out.