Remove empty cell for state of branding
For some reason the cell used to hold the branding
text causes an error with the JAWS screen reader
when empty (which it is when the 'only branding'
brand has no text).

In this state, and when the content of the email
has a certain form, JAWS announces all links in
the email the same.

This therefore removes that cell when it is empty.

It includes the addition of a max-width to stop
the Outlook app on Android from stretching the
logo to twice the size. It seems strange to keep
the height attribute on the <img> tag but that is
needed to keep the image working on Outlook on
Windows.

Also note that we only apply the max-width when
there is no text. Setting it on the cell with the
logo image makes it overlap the text in several
email clients.