Lowercase UUIDs in cache keys
When we cache things in Redis the keys are case sensitive.

Our keys in Redis are usually comprised of:
- hard coded strings which we control (for example `"templates"` or
  `"service"`)
- UUIDs
- integers (for example template version `99`)
- `None`

Our keys were not dealing with arbitrary user-provided strings, so we
didn’t think we needed to worry about case sensitivity.

However our ORM and database accept UUIDs in upper, lower or mixed case.
This means it was possible to:
- cache an object with key of `item-ABC123`
- clear the cache for `item-abc123` (for example by modifying or
  deleting the object)
- fetch the object with the key `item-ABC123`, thereby accesing a stale
  cache

This commit adds some pre-processing of the cache keys to lowercase any
argument value of the underlying function which looks like a UUID.

This means that when we are caching things in Redis UUIDs in cache keys
will always be lowercase, and therefore we will only have one cache
entry for a given object.

This commit does not lowercase all strings. Doing so would likely cause
short-term bugs doing deployment where – for example – some parts of the
code would be using a cache key of `version-None` and others
`version-none`.