Remove translation of markdown links in letters
When we first added markdown-style formatting to emails we – as a
convenience – decided to translate links into a text-only-friendly
representation, so `[link](url)` became `link: url`:

https://github.com/alphagov/notifications-utils/blob/13abc3f200603c46c35c76b0dc40718c7efe3ac2/notifications_utils/formatters.py#L149-L150

This is because we did not – officially or unofficially – support link
text in emails.

When we added letter templates we shared the same method for rendering
links, so letter templates also got this same plain-text representation
of Markdown-style links.

The idea was that if you pasted a Markdown-style link into a template
and didn’t realise it wasn’t supported the recipient wouldn’t get
something which looked broken.

We later chose to make the URLs bold, the same as other URLs in letters:
https://github.com/alphagov/notifications-utils/commit/713c765a916218f0b826915876e0bbddf66e1b52

In January 2018 we enabled link text in emails, but didn’t offically
support it: https://github.com/alphagov/notifications-utils/pull/293/files

In August 2023 we officially documented support for link text in emails:
https://github.com/alphagov/notifications-admin/commit/da6ad0d01866df56c03ebe0572e4e6d4a4b3e58f

So the fact that link text works in letters is a bit of a historic
oddity. Doing some analysis shows that only a couple of real templates
are using this behaviour, and probably by accident.

Removing support for this will make our code easier to maintain, because
it means we have:
- one less undocumented feature that we need to maintain support for
- one less bit of weirdness that people need to spelunk through the code
  for (see above) in order to understand

We were bitten by both of the above when working on QR codes in letters.

Before we deply this change we will need to manually edit the couple of
templates which are relying on it