Add 'hidden' attribute to email preheader
This was requested by a user who questioned our
use of a 'preheader' paragraph at the start of
emails (to control the summary shown for that
email by the email client) hidden by CSS. They
stated that some email clients allow HTML but
without CSS so it can appear. Use of the 'hidden'
attribute will solve this.

Note: this change is actually the main reason
we're changing the doctype to HTML5