Lift and shift LetterImageTemplate
Moves this class and its related template file to the admin app, which
is the only thing using it.

This allow us to inject overlay (UI) elements into the correct place, as
we want to have an 'edit' button appear on the first page of English
content. This may be any arbitrary depth into the letter.