Allow hiding notify tag on letter print template
This tag should only be present on the first page of a PDF we generate
to send to DVLA.

In supporting bilingual letters, we generate two PDFs: a Welsh variant
followed by an English variant.

We need to stop inserting the 'NOTIFY' tag on the English variant, and
so need to be able to toggle the tag off, otherwise DVLA's printers get
confused.