Make it clearer what is being tested
It shouldn’t be possible to create a `SerialisedModel` instance where
an `@property` has the same name as a field in `ALLOWED_PROPERTIES`.

This test makes sure that an exception is raised if anyone tries to do
this (rather than a silent failure or some other unexpected behaviour).

We don’t need to pass in or read any data for the exception to be raised
so this commit removes the test code which is doing that and obscuring
the purpose of the test.