logging: add microseconds to time format via mixin class
ideally we'd just include this in our TIME_FORMAT, but python's
strftime has no ability to include even milliseconds. the
python-default logging datefmt is augmented with milliseconds
in the stdlib logging module via a hack, but this hack is disabled
when a custom datefmt is supplied. so we have to add back our own
hack to include precise time. use microsecond granularity as it
should be enough to correctly order even log records that are
generated immediately-consecutively in the same thread.