NotifyRequest: handle trace ids when not handed X-B3-TraceId by paas
firstly look for alternate headers NOTIFY_TRACE_ID_ALT_HEADERS
(default cloudfront's and the ALB's), then failing that generate
a random trace id.

add similar configurability for span id headers and parent span
id headers - this makes the code more symmetrical, even if we
don't use the configuration options now.

add a get_onwards_request_headers method which can be used by
internal api clients etc. when preparing to make onwards
requests. it will automatically handle the advance of span_id
and parent_span_id correctly.

FWIW taking a cloudfront or ALB request id and passing it
around as a `X-B3-TraceId` is strictly not zipkin compliant
(because zipkin nowdays stipulates it wants a hexadecimal
identifer of a certain number of characters). Ah well.

almost all of this code is from digitalmarketplace-utils,
including the extensive tests. omitted was support for sample
and debug flag handling, because presumably that will be handled
by sentry and/or opentelemetry at some point. that said, it's
possible we'll rip all this out in favour of a generic
opentelemetry solution in the future.