NotifyRequest: generate own span_id if none provided in headers
the worry about such a span_id being propagated indefinitely is
partially valid, but on balance it's probably worse not having *any*
span id for the initial span as it makes it harder to find logs and
failing to pass on a parent span id to the next server is
indistinguishable from the mechanism being broken. and on a platform
that isn't using a span id-aware router between apps, we should just
handle the advancing of span ids ourselves.

when generating own span_id, prefix it with `self-` in attempt to
avoid people going looking for whatever caller assigned it this
span_id. this is technically not a zipkin-compliant value, but we're
already breaking those rules by passing around cloudfront and ALB
request ids under the x-b3-traceid header. this should only ever get
passed in requests under the x-b3-parentspanid header.