add gunicorn_defaults module
this contains a lot of common configuration we apply (or *should*
apply) across all of our gunicorn servers. implemented as a
globals-mutating function to allow it to make mutating changes
to things beyond its module. indeed it is weird and a little bit
ugly, but I think it's inevitable.

includes new dict-based config for gunicorn's logging which will
result in structured logs.