Preserve the first, not last item found
Normal `dict`s can’t have duplicate keys. `InsensitiveDict`s can receive
duplicate items, where normalising two different keys would cause them
to be the same.

At the moment the default behaviour is to keep the last-seen value.

This commit adds a flag to preserve the first-seen value, and not
overwrite it.

This is most relevant for how we’re going to use `InsensitiveSet`, where
it makes more sense for the format of a placeholder shown to the user to
be how it first occurs in a template.