Enable pyupgrade linting rules
Pyupgrade[1] automatically upgrades syntax for newer versions of Python.
Newer syntax is often friendlier and less verbose – for example f-string
formatting rather than calls to `str.format`.

It has a total of 42 rules[2].

This commit adds it to our linter config, and then fixes all the
instances of old syntax it found (the vaast majority of these were
automatic by running Ruff with the `--fix` flag).

1. https://pypi.org/project/pyupgrade/
2. https://docs.astral.sh/ruff/rules/#pyupgrade-up