Re-enable flake8-print
We don’t want print statements in production code – should be using
logging instead.

Ruff doesn’t have this check on by default[1] so we need to explicitly
include it.

Some existing print statements in this codebase are there intentionally
(they are in a command not a public endpoint) so I’ve marked them as
exempt from checking by running `ruff check <file> --add-noqa`

Adding this now so our linting rules have parity with the admin app,
which is the first app we’re going to copy the utils version of
`pyproject.toml` into.

1. https://docs.astral.sh/ruff/rules/print/