Try second redis delete when updating API resource to avoid race
In the previous commit, we tried to avoid stale data in redis if
redis is unavailable. However, this introduced a race condition
where another request re-populates the redis key with the old value
before the database has been updated. We fix this by also deleting
the cache key after the API update so that if that happened, we would
remove the stale data after the API update.

There is a small edge case that could still end up with stale data
that this commit doesn't solve (and we may just have to tolerate).
If we hit the race condition above so stale data is reinserted in
redis before the database update, then the database update happens
and redis goes down at this point. Then what will happen is the
second cache clear will fail and we will be left with stale data.
This should be a rare case as involves both a race condition happening
so two requests for the same object at very similar times, and redis
going down mid request.

Again, we lean towards throwing an uncaught exception if a redis
delete fails. It means there is a chance of stale data and the
user should hopefully check if their action was successful and
retry if they see stale data in the user interface (which should
then hopefully fix the problem).